chore: added seperat reference file for livemix
